added role hierarchy throughout the admin console.  can only access i…

+++ checklist/src/components/ChangePrivilegeList.js
@@ -2,15 +2,17 @@ import React, { Component } from "react";
 import { Button } from "antd";
 import ChangePrivilegeListItem from "./ChangePrivilegeListItem.js";
 import firebase from "../configs/firebaseConfig";
+import roleHierarchy from "../roles/roleHierarchy.js";
 
 /* PROPS:
     firebasePath: string; the url the data is being pulled from to display (i.e. /users)
     fieldsToDisplay: [obj]; array of objects representing the fields and their prompts, as such
       field: "name",
       visibleDescription: "User's Name:"
-    arrayOfPrivileges: [string]; array of the possible strings "privilege" could be set to
+    arrayOfPrivileges: [string]; array of the possible strings "role" could be set to
       (i.e. ["intern", "admin"])
-      IMPORTANT: this file makes the assumption the field we wish to edit is called "privilege"
+      IMPORTANT: this file makes the assumption the field we wish to edit is called "role"
+    userInfo: object; details the currently logged-in user
 */
 
 /* STATE:
@@ -80,12 +82,19 @@ class ChangePrivilegeList extends Component {
   }
 
   render() {
+    // the roles that the logged-in user is allowed to see based on the role hierarchy
+    const rolesUserCanSee = roleHierarchy[this.props.userInfo.role];
+
     // map through the firebase data and create the ListItems to be rendered
     const listObjects = this.state.firebaseData
       .sort((a, b) => {
         return a.location > b.location;
       })
       .map(dataObject => {
+        // only return the list item if the logged-in user can see that role
+        if (!rolesUserCanSee.includes(dataObject.role)) {
+          return <div />;
+        }
         return (
           <ChangePrivilegeListItem
             onDelete={() => this.onDelete(dataObject)}
